OT166-94: Arreglos PR en ActivityIntegationTest, BigTest, pom(nueva dependencia)
