+++ pom.xml
@@ -172,6 +172,11 @@
       <artifactId>springdoc-openapi-ui</artifactId>
       <version>1.6.6</version>
     </dependency>
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-lang3</artifactId>
+      <version>3.12.0</version>
+    </dependency>
   </dependencies>
 
   <description>ONG project with Spring Boot</description>
